hopefully removed all the remaining noisy messages from gcc
git-svn-id: https://svn.kvirc.de/svn/trunk/kvirc@2841 17fca916-40b9-46aa-a4ea-0a15b648b75c

+++ src/kvirc/kvs/kvi_kvs_variant.cpp
@@ -630,7 +630,7 @@ void KviKvsVariant::dump(const char * prefix) const
 		case KviKvsVariantData::String: debug("%s String(%s) [this=0x%lx]",prefix,m_pData->m_u.pString->toUtf8().data(),(long unsigned int)this); break;
 		case KviKvsVariantData::Array: debug("%s Array(ptr=0x%lx) [this=0x%lx]",prefix,(long unsigned int)m_pData->m_u.pArray,(long unsigned int)this); break;
 		case KviKvsVariantData::Hash: debug("%s Hash(ptr=0x%lx,dict=0x%lx) [this=0x%lx]",prefix,(long unsigned int)m_pData->m_u.pHash,(long unsigned int)m_pData->m_u.pHash->dict(),(long unsigned int)this); break;
-		case KviKvsVariantData::Integer:  debug("%s Integer(%d) [this=0x%lx]",prefix,m_pData->m_u.iInteger,(long unsigned int)this); break;
+		case KviKvsVariantData::Integer:  debug("%s Integer(%d) [this=0x%lx]",prefix,(int) m_pData->m_u.iInteger,(long unsigned int)this); break;
 		case KviKvsVariantData::Real: debug("%s Real(%f) [this=0x%lx]",prefix,*(m_pData->m_u.pReal),(long unsigned int)this); break;
 		case KviKvsVariantData::Boolean: debug("%s Boolean(%s) [this=0x%lx]",prefix,m_pData->m_u.bBoolean ? "true" : "false",(long unsigned int)this); break;
 		case KviKvsVariantData::HObject: debug("%s HObject(%lx) [this=0x%lx]",prefix,(long unsigned int)m_pData->m_u.hObject,(long unsigned int)this); break;
